Kubernetes Power Manager upversion
This commit applies the necessary changes to upversion the
Kubernetes Power Manager app, from v2.3.0 to v2.4.0.

The version v2.4.0 incorporates the proposed changes to the
upstream, what allows the removal of the image build script
and patches, as well the use of public docker image.

This change also proposes the removal of the local helm code,
giving preference to the use of the code available on the
upstream (published with v2.4.0). In this new approach, the
helm code is incorporated and adapted by applying patches.

Test Plan:
PASS: AIO-SX: check  if  the  application  can be successfully
      installed.
PASS: check if label "app.starlingx.io/component" was  applied
      in "intel-power" namespace on install.
PASS: check platform cores allocation
PASS: AIO-SX: check  if  the  application  can be successfully
      removed and deleted.
PASS: check if installed CRDs are removed on application
      removal/deletion.
PASS: check if daemonsets (agents) started  by  the controller
      are removed on application removal/deletion.
PASS: check if the namespace was deleted on application
      removal/deletion.
PASS: check if the app is correctly updated from earlier version.

Story: 2011069
Task: 49757
Depends-On: https://review.opendev.org/c/starlingx/config/+/913696

Author: Eduardo Alberti <eduardo.alberti@windriver.com>
Co-author: Guilherme Henrique Pereira dos Santos <guilherme.santos@windriver.com>

Change-Id: Ic701cf29e9d6650f0b832960efa343f09e43adc8
Signed-off-by: Eduardo Juliano Alberti <eduardo.alberti@windriver.com>