Kubernetes Power Manager Helm Charts upversion to v2.5.0
This change aims to update the patch applied to the upstream Helm Chart,
adapting it to the new version of Kubernetes Power Manager (v2.5.0).

Story: 2011069
Task: 50276

Test Plan
PASS: AIO-SX: check if the application can be successfully
      installed.
PASS: AIO-SX: check if the application can be successfully
      uninstalled.
PASS: check if installed CRDs are correctly created.

Change-Id: Idc2f22ccbca27a47f6b19ed3f5a3faffe111bf47
Signed-off-by: Eduardo Alberti <eduardo.alberti@windriver.com>