Delegates dependency validation to app framework
Considering the available actions for dependency validation in the app
framework, the override to allow the power manager installation is not
necessary anymore.
This commit removes the dependency validation from python plugin and
delegates the control to the app framework.

TEST PLAN:
PASS: APP installed successfully without NFD
PASS: Warning message shown when NFD is not installed

Story: 2011624
Task: 53312

Change-Id: I3fa7f237f5008efea936b1d394c7c2b340908001
Signed-off-by: Eduardo Alberti <eduardo.alberti@windriver.com>